- Fix GPS Status Switching
- Rename BLE Advertisement
- Change Battery Threshold

+++ include/common.h
@@ -9,8 +9,8 @@
 
 struct GPSData_t
 {
-    double longitude;
-    double latitude;
+    float longitude;
+    float latitude;
     char status;
 };
